Remove generated top level module.
